Editors entry form: organizations as editors
