+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/edit/n3editing/configuration/generators/AddEditorsToInformationResourceGenerator.java
@@ -28,7 +28,7 @@ import org.apache.jena.vocabulary.XSD;
 
 import edu.cornell.mannlib.vitro.webapp.controller.VitroRequest;
 import edu.cornell.mannlib.vitro.webapp.dao.jena.QueryUtils;
-import edu.cornell.mannlib.vitro.webapp.edit.n3editing.FirstAndLastNameValidator;
+import edu.cornell.mannlib.vitro.webapp.edit.n3editing.PublicationHasAuthorValidator;
 import edu.cornell.mannlib.vitro.webapp.edit.n3editing.VTwo.EditConfigurationUtils;
 import edu.cornell.mannlib.vitro.webapp.edit.n3editing.VTwo.EditConfigurationVTwo;
 import edu.cornell.mannlib.vitro.webapp.edit.n3editing.VTwo.fields.FieldVTwo;
@@ -67,6 +67,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 
         editConfiguration.addNewResource("editorshipUri", DEFAULT_NS_TOKEN);
         editConfiguration.addNewResource("newPerson", DEFAULT_NS_TOKEN);
+	editConfiguration.addNewResource("newOrg", DEFAULT_NS_TOKEN);
         editConfiguration.addNewResource("vcardPerson", DEFAULT_NS_TOKEN);
         editConfiguration.addNewResource("vcardName", DEFAULT_NS_TOKEN);
 
@@ -85,7 +86,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
         //template file
         editConfiguration.setTemplate("addEditorsToInformationResource.ftl");
         //add validators
-        editConfiguration.addValidator(new FirstAndLastNameValidator("personUri", I18n.bundle(vreq)));
+        editConfiguration.addValidator(new PublicationHasAuthorValidator());
 
         //Adding additional data, specifically edit mode
         addFormSpecificData(editConfiguration, vreq);
@@ -126,7 +127,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 
 	private String getN3EditorshipRank() {
 		return getN3PrefixString() +
-        "?editorshipUri core:editorRank ?rank .";
+        "?editorshipUri core:rank ?rank .";
 	}
 
 	//first name, middle name, last name, and new perseon for new editor being created, and n3 for existing person
@@ -144,7 +145,9 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
                 getN3NewVCardPerson(),
 
                 getN3EditorshipRank(),
-                getN3ForExistingPerson());
+                getN3ForExistingPerson(),
+		getN3NewOrg(),
+                getN3ForExistingOrg());
 	}
 
 
@@ -219,10 +222,10 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 
 	// Changes here for creating vcards for external editors
 	private String getN3NewVCardPerson() {
-		return  getN3PrefixString() + getDisableVCardPrefix() +
-		"@prefix vcard: <http://www.w3.org/2006/vcard/ns#> .  \n" +
-		"?vcardPerson a vcard:Individual ;\n" +
-		"<" + RDFS.label.getURI() + "> ?label .\n" +
+	return  getN3PrefixString() + getDisableVCardPrefix() +
+	"@prefix vcard: <http://www.w3.org/2006/vcard/ns#> .  \n" +
+	"?vcardPerson a vcard:Individual ;\n" +
+	"<" + RDFS.label.getURI() + "> ?label .\n" +
         "?editorshipUri core:relates ?vcardPerson .\n" +
         "?vcardPerson core:relatedBy ?editorshipUri . ";
 	}
@@ -231,6 +234,20 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 		return getN3PrefixString() +
 		"?editorshipUri core:relates ?personUri .\n" +
 		"?personUri core:relatedBy ?editorshipUri .";
+	}
+
+	private String getN3NewOrg() {
+		return  getN3PrefixString() +
+        "?newOrg a foaf:Organization ;\n" +
+        "<" + RDFS.label.getURI() + "> ?orgName .\n" +
+        "?editorshipUri core:relates ?newOrg .\n" +
+        "?newOrg core:relatedBy ?editorshipUri . ";
+	}
+
+	private String getN3ForExistingOrg() {
+		return getN3PrefixString() +
+		"?editorshipUri core:relates ?orgUri .\n" +
+		"?orgUri core:relatedBy ?editorshipUri .";
 	}
 
 	/**  Get new resources	 */
@@ -244,6 +261,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 			newResources.put("newPerson", DEFAULT_NS_TOKEN);
 			newResources.put("vcardPerson", DEFAULT_NS_TOKEN);
 			newResources.put("vcardName", DEFAULT_NS_TOKEN);
+			newResources.put("newOrg", DEFAULT_NS_TOKEN);
 			return newResources;
 		}
 
@@ -266,6 +284,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 
     	urisOnForm.add("createPersonInstance");
     	urisOnForm.add("createVCard");
+    	urisOnForm.add("orgUri");
     	editConfiguration.setUrisOnform(urisOnForm);
 
     	//for person who is not in system, need to add first name, last name and middle name
@@ -274,6 +293,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
     			"middleName",
     			"lastName",
     			"rank",
+			"orgName",
     			"label");
     	editConfiguration.setLiteralsOnForm(literalsOnForm);
     }
@@ -303,6 +323,8 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
     	setRankField(editConfiguration);
     	setPersonUriField(editConfiguration);
     	setPatternFailFields(editConfiguration);
+    	setOrgUriField(editConfiguration);
+    	setOrgNameField(editConfiguration);
     }
 
     private void setPatternFailFields(EditConfigurationVTwo editConfiguration) {
@@ -360,6 +382,20 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 				);
 	}
 
+	private void setOrgUriField(EditConfigurationVTwo editConfiguration) {
+		editConfiguration.addField(new FieldVTwo().
+				setName("orgUri")
+				);
+	}
+
+	private void setOrgNameField(EditConfigurationVTwo editConfiguration) {
+		editConfiguration.addField(new FieldVTwo().
+				setName("orgName").
+				setValidators(list("datatype:" + XSD.xstring.toString())).
+				setRangeDatatypeUri(XSD.xstring.toString())
+				);
+	}
+
 	//Form specific data
 	public void addFormSpecificData(EditConfigurationVTwo editConfiguration, VitroRequest vreq) {
 		HashMap<String, Object> formSpecificData = new HashMap<String, Object>();
@@ -394,7 +430,7 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 			+ "        ?subject core:relatedBy ?editorshipURI .\n"
 			+ "        ?editorshipURI a core:Editorship .\n"
 			+ "        ?editorshipURI core:relates ?editorURI .\n"
-			+ "        ?editorURI a foaf:Person .\n"
+			+ "        ?editorURI a foaf:Agent .\n"
 			+ "        ?editorURI a ?type .\n"
 			+ "    }\n"
 			+ "    UNION\n"
@@ -402,14 +438,14 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
 			+ "        ?subject core:relatedBy ?editorshipURI .\n"
 			+ "        ?editorshipURI a core:Editorship .\n"
 			+ "        ?editorshipURI core:relates ?editorURI .\n"
-			+ "        ?editorURI a foaf:Person .\n"
-			+ "        ?editorURI rdfs:label ?editorName\n"
+			+ "        ?editorURI a foaf:Agent .\n"
+			+ "        ?editorURI rdfs:label ?editorName .\n"
 			+ "    }\n"
 			+ "    UNION\n"
 			+ "    {\n"
 			+ "        ?subject core:relatedBy ?editorshipURI .\n"
 			+ "        ?editorshipURI a core:Editorship .\n"
-			+ "        ?editorshipURI core:rank ?rank\n"
+			+ "        ?editorshipURI core:rank ?rank .\n"
 			+ "    }\n"
 			+ "    UNION\n"
 			+ "    {\n"
@@ -438,13 +474,13 @@ public class AddEditorsToInformationResourceGenerator extends VivoBaseGenerator 
         + "PREFIX core: <http://vivoweb.org/ontology/core#> \n"
         + "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#> \n"
         + "PREFIX foaf: <http://xmlns.com/foaf/0.1/> \n"
-		+ "PREFIX vcard:  <http://www.w3.org/2006/vcard/ns#> \n"
+	+ "PREFIX vcard: <http://www.w3.org/2006/vcard/ns#> \n"
         + "SELECT ?editorshipURI (REPLACE(STR(?editorshipURI),\"^.*(#)(.*)$\", \"$2\") AS ?editorshipName) ?editorURI ?editorName ?rank \n"
         + "WHERE { { \n"
         + "?subject core:relatedBy ?editorshipURI . \n"
         + "?editorshipURI a core:Editorship . \n"
         + "?editorshipURI core:relates ?editorURI . \n"
-        + "?editorURI a foaf:Person . \n"
+        + "?editorURI a foaf:Agent . \n"
         + "OPTIONAL { ?editorURI rdfs:label ?editorName } \n"
         + "OPTIONAL { ?editorshipURI core:rank ?rank } \n"
 		+ "} UNION {  \n"
