+++ webapp/src/main/webapp/templates/freemarker/edit/forms/addEditorsToInformationResource.ftl
@@ -18,6 +18,7 @@
 <#assign lastNameValue = lvf.getFormFieldValue(editSubmission, editConfiguration, "lastName") />
 <#assign firstNameValue = lvf.getFormFieldValue(editSubmission, editConfiguration, "firstName") />
 <#assign middleNameValue = lvf.getFormFieldValue(editSubmission, editConfiguration, "middleName") />
+<#assign orgNameValue = lvf.getFormFieldValue(editSubmission, editConfiguration, "orgName") />
 
 <#--UL class based on size of existing editors-->
 <#assign ulClass = ""/>
@@ -101,6 +102,13 @@
 <form id="addEditorForm" action ="${submitUrl}" class="customForm noIE67">
     <h3>${i18n().add_an_editor}</h3>
 
+    <div style="display:inline">
+        <input type="radio" name="editorType" class="person-radio" value="" role="radio" checked />
+        <label class="inline" for="Person" >${i18n().person_capitalized}</label>
+        <input type="radio" name="editorType" class="org-radio" value="http://xmlns.com/foaf/0.1/Organization" role="radio" style="display:inline;margin-left:18px" />
+        <label class="inline" for="Organization">${i18n().organization_capitalized}</label>
+    </div>
+
     <section id="personFields" role="personContainer">
     		<#--These wrapper paragraph elements are important because javascript hides parent of these fields, since last name
     		should be visible even when first name/middle name are not, the parents should be separate for each field-->
@@ -137,6 +145,22 @@
         </div>
     </section>
 
+    <section id="organizationFields" role="organization">
+    		<p class="inline">
+        <label for="orgName">${i18n().organization_name_capitalized} <span class='requiredHint'> *</span></label>
+        <input size="38"  type="text" id="orgName" name="orgName" value="${orgNameValue}" role="input" />
+        </p>
+
+        <div id="selectedOrg" class="acSelection">
+            <p class="inline">
+                <label>${i18n().selected_organization}:&nbsp;</label>
+                <span  id="selectedOrgName"></span>
+                <a href="${urls.base}/individual?uri=" id="orgLink"  title="${i18n().verify_match_capitalized}">(${i18n().verify_match_capitalized})</a>
+                <input type="hidden" id="orgUri" name="orgUri" value=""  role="input" /> <!-- Field value populated by JavaScript -->
+            </p>
+        </div>
+    </section>
+
     <input type="hidden" id="label" name="label" value=""  role="input" />  <!-- Field value populated by JavaScript -->
 
 
@@ -161,6 +185,7 @@ var customFormData = {
     acUrl: '${urls.base}/autocomplete?type=',
     tokenize: '&tokenize=true',
     personUrl: 'http://xmlns.com/foaf/0.1/Person',
+    orgUrl: 'http://xmlns.com/foaf/0.1/Organization',
     reorderUrl: '${urls.base}/edit/reorder'
 };
 var i18nStrings = {
@@ -169,6 +194,7 @@ var i18nStrings = {
     removeEditorshipMessage: '${i18n().confirm_editor_removal?js_string}',
     removeEditorshipAlert: '${i18n().error_processing_editor_request?js_string}',
     editorTypeText: '${i18n().editor_capitalized?js_string}',
+    organizationTypeText: '${i18n().organization_capitalized?js_string}',
     helpTextSelect: '${i18n().select_an_existing?js_string}',
     helpTextAdd: '${i18n().or_add_new_one?js_string}'
 };
