+++ webapp/src/main/webapp/templates/freemarker/edit/forms/js/addEditorsToInformationResource.js
@@ -64,9 +64,20 @@ var addEditorForm = {
         this.selectedEditorName = $('#selectedEditorName');
         this.acHelpTextClass = 'acSelectorWithHelpText';
         this.verifyMatch = this.form.find('.verifyMatch');
+        this.personRadio = $('input.person-radio');
+        this.orgRadio = $('input.org-radio');
         this.personSection = $('section#personFields');
+        this.orgSection = $('section#organizationFields');
+        this.orgName = $('input#orgName');
+        this.orgNameWrapper = this.orgName.parent();
+        this.orgUriField = $('input#orgUri');
+        this.selectedOrg = $('div#selectedOrg');
+        this.selectedOrgName = $('span#selectedOrgName');
+        this.orgLink = $('a#orgLink');
         this.personLink = $('a#personLink');
         this.returnLink = $('a#returnLink');
+		
+		this.orgSection.hide();
     },
 
     // Initial page setup. Called only at page load.
@@ -158,6 +169,8 @@ var addEditorForm = {
 
         this.hideSelectedPerson();
 
+        addEditorForm.setEditorType("person");
+
         this.cancel.unbind('click');
         this.cancel.bind('click', function() {
             addEditorForm.showEditorListOnlyView();
@@ -181,14 +194,22 @@ var addEditorForm = {
         this.personUriField.val('');
     },
 
+    hideSelectedOrg: function() {
+        this.selectedOrg.hide();
+        this.selectedOrgName.html('');
+        this.orgUriField.val('');
+    },
+
     showFieldsForNewPerson: function() {
         this.firstNameWrapper.show();
         this.middleNameWrapper.show();
+        $('#createVCard').parent().show();
     },
 
     hideFieldsForNewPerson: function() {
         this.hideFields(this.firstNameWrapper);
         this.hideFields(this.middleNameWrapper);
+        $('#createVCard').parent().hide();
     },
 
     /* *** Ajax initializations *** */
@@ -200,10 +221,20 @@ var addEditorForm = {
         // an editor.
         this.acCache = {};
         this.setAcFilter();
-        var $acField = this.lastNameField;
-        var urlString = addEditorForm.acUrl + addEditorForm.personUrl + addEditorForm.tokenize;
-        var authType = "person";
+        var $acField;
+        var urlString;
+        var editorType;
 
+        if  ( this.personRadio.prop("checked") ) {
+            $acField = this.lastNameField;
+            urlString = addEditorForm.acUrl + addEditorForm.personUrl + addEditorForm.tokenize;
+            editorType = "person";
+        }
+        else {
+            $acField = this.orgName;
+            urlString = addEditorForm.acUrl + addEditorForm.orgUrl + addEditorForm.tokenize;
+            editorType = "org";
+        }
         $acField.autocomplete({
             minLength: 2,
             source: function(request, response) {
@@ -241,7 +272,7 @@ var addEditorForm = {
             // select event has been triggered. To trigger a select, the user must hit enter
             // or click on the selection with the mouse. This appears to confuse some users.
             select: function(event, ui) {
-                addEditorForm.showSelectedEditor(ui,authType);
+                addEditorForm.showSelectedEditor(ui,editorType);
             }
         });
 
@@ -297,9 +328,9 @@ var addEditorForm = {
     },
 
     // Action taken after selecting an editor from the autocomplete list
-    showSelectedEditor: function(ui,authType) {
+    showSelectedEditor: function(ui,editorType) {
 
-        if ( authType == "person" ) {
+        if ( editorType == "person" ) {
             this.personUriField.val(ui.item.uri);
             this.selectedEditor.show();
 
@@ -315,12 +346,23 @@ var addEditorForm = {
             this.hideFieldsForNewPerson();
             this.personLink.attr('href', this.verifyMatch.data('baseHref') + ui.item.uri);
         }
+        else {
+            // do the same as above but for the organization fields
+            this.orgUriField.val(ui.item.uri);
+            this.selectedOrg.show();
+
+            this.selectedOrgName.html(ui.item.label);
+
+            this.hideFields(this.orgNameWrapper);
+
+            this.orgLink.attr('href', this.verifyMatch.data('baseHref') + ui.item.uri);
+        }
 
         // Cancel restores initial form view
         this.cancel.unbind('click');
         this.cancel.bind('click', function() {
             addEditorForm.initFormView();
-            addEditorForm.setEditorType(authType);
+            addEditorForm.setEditorType(editorType);
             return false;
         });
     },
@@ -444,6 +486,14 @@ var addEditorForm = {
             return false;
         });
 
+        this.orgRadio.click(function() {
+            addEditorForm.setEditorType("org");
+        });
+
+        this.personRadio.click(function() {
+            addEditorForm.setEditorType("person");
+        });
+
         this.form.submit(function() {
             // NB Important JavaScript scope issue: if we call it this way, this = addEditorForm
             // in prepareSubmit. If we do this.form.submit(this.prepareSubmit); then
@@ -469,6 +519,11 @@ var addEditorForm = {
             return false;
         });
 
+        this.orgLink.click(function() {
+            window.open($(this).attr('href'), 'verifyMatchWindow', 'width=640,height=640,scrollbars=yes,resizable=yes,status=yes,toolbar=no,menubar=no,location=no');
+            return false;
+        });
+
     	this.acSelector.focus(function() {
         	addEditorForm.deleteAcHelpText(this);
     	});
@@ -476,6 +531,15 @@ var addEditorForm = {
     	this.acSelector.blur(function() {
         	addEditorForm.addAcHelpText(this);
     	});
+
+        this.orgName.focus(function() {
+        	addEditorForm.deleteAcHelpText(this);
+    	});
+
+    	this.orgName.blur(function() {
+        	addEditorForm.addAcHelpText(this);
+    	});
+
 
         // When hitting enter in last name field, show first and middle name fields.
         // NB This event fires when selecting an autocomplete suggestion with the enter
@@ -501,7 +565,7 @@ var addEditorForm = {
             name;
 
         // If selecting an existing person, don't submit name fields
-        if (this.personUriField.val() != '' ) {
+        if (this.personUriField.val() != '' || this.orgUriField.val() != '' || this.orgName.val() != '' ) {
             this.firstNameField.attr('disabled', 'disabled');
             this.middleNameField.attr('disabled', 'disabled');
             this.lastNameField.attr('disabled', 'disabled');
@@ -521,6 +585,10 @@ var addEditorForm = {
 
             this.labelField.val(name);
         }
+		// If user selected org via autocomplete, clear the org name field
+		if ( this.orgUriField.val() != '' ) {
+			this.orgName.val("");
+		}
 
     },
 
@@ -663,6 +731,9 @@ var addEditorForm = {
         if ( $(selectedObj).attr('id') == "lastName" ) {
             typeText = addEditorForm.editorTypeText;
         }
+        else {
+            typeText = addEditorForm.organizationTypeText;
+        }
 
         if (!$(selectedObj).val()) {
 			$(selectedObj).val(addEditorForm.helpTextSelect + " " + typeText + " " + addEditorForm.helpTextAdd)
@@ -679,9 +750,54 @@ var addEditorForm = {
 
     // we need to set the correct class names for fields like the acSelector, acSelection, etc.
     // as well as clear and disable fields, call other functions ...
-	setEditorType: function(authType) {
-        if ( authType == "person" ) {
+	setEditorType: function(editorType) {
+        if ( editorType == "org" ) {
+	        this.personSection.hide();
+	        this.orgSection.show();
+			this.orgNameWrapper.show();
+	        // person fields
+            this.personRadio.prop('checked', false);  // needed for reset when cancel button is clicked
+	        this.acSelector.removeClass("acSelector");
+	        this.acSelector.removeClass(this.acHelpTextClass);
+	        this.selectedEditor.removeClass("acSelection");
+	        this.selectedEditorName.removeClass("acSelectionInfo");
+	        this.personLink.removeClass("verifyMatch");
+	        this.acSelector.attr('disabled', 'disabled');
+	        this.firstNameField.attr('disabled', 'disabled');
+	        this.middleNameField.attr('disabled', 'disabled');
+	        this.lastNameField.attr('disabled', 'disabled');
+	        this.acSelector.val('');
+	        this.firstNameField.val('');
+	        this.middleNameField.val('');
+	        this.lastNameField.val('');
+	        // org fields
+	        this.orgRadio.prop('checked', true); // needed for reset when cancel button is clicked
+	        this.orgName.addClass("acSelector");
+	        this.selectedOrg.addClass("acSelection");
+	        this.selectedOrgName.addClass("acSelectionInfo");
+	        this.orgLink.addClass("verifyMatch");
+	        this.orgName.attr('disabled', false);
+	        this.orgUriField.attr('disabled', false);
+
+	        addEditorForm.addAcHelpText(this.orgName);
+	        addEditorForm.initAutocomplete();
+	        addEditorForm.hideSelectedPerson();
+	    }
+	    else if ( editorType == "person" ) {
+	        this.orgSection.hide();
 	        this.personSection.show();
+	        // org fields
+	        this.orgRadio.prop('checked', false);  // needed for reset when cancel button is clicked
+	        this.orgName.removeClass("acSelector");
+	        this.orgName.removeClass(this.acHelpTextClass);
+	        this.selectedOrg.removeClass("acSelection");
+	        this.selectedOrgName.removeClass("acSelectionInfo");
+	        this.orgLink.removeClass("verifyMatch");
+	        this.orgName.attr('disabled', 'disabled');
+	        this.orgUriField.attr('disabled', 'disabled');
+	        this.orgName.val('');
+	        this.orgUriField.val('');
+            // person fields
             this.acSelector.addClass("acSelector");
             this.personRadio.prop('checked', true);  // needed for reset when cancel button is clicked
 	        this.selectedEditor.addClass("acSelection");
@@ -694,6 +810,8 @@ var addEditorForm = {
 
 	        addEditorForm.addAcHelpText(this.acSelector);
 	        addEditorForm.initAutocomplete();
+	        addEditorForm.hideSelectedOrg();
+
 	    }
     }
 };
